+++ hotel-member/src/main/java/sustech/hotel/member/entity/CollectHotelEntity.java
@@ -14,5 +14,5 @@ public class CollectHotelEntity implements Serializable {
     private static final long serialVersionUID = 1L;
 
     private Long userId;
-    private Integer hotelId;
+    private Long hotelId;
 }
